Move all Timestamp-related into separate file
