+++ lib/encoder.js
@@ -4,10 +4,6 @@ var Buffer = require('safe-buffer').Buffer
 var bl = require('bl')
 
 module.exports = function buildEncode (encodingTypes, options) {
-  if (!options.disableTimestampEncoding) {
-    encodingTypes.push(encoderDate)
-  }
-
   function encode (obj) {
     if (obj === undefined) throw new Error('undefined is not encodable in msgpack!')
     if (isNaN(obj)) throw new Error('NaN is not encodable in msgpack!')
@@ -143,15 +139,6 @@ function encodeFloat (obj, forceFloat64) {
   return buf
 }
 
-const encoderDate = {
-  check (obj) {
-    return typeof obj.getDate === 'function'
-  },
-  encode (obj) {
-    return encodeDate(obj)
-  }
-}
-
 function encodeExt (obj, encodingTypes) {
   var encoded
 
@@ -258,33 +245,6 @@ function getBufferHeader (length) {
   }
 
   return header
-}
-
-function encodeDate (dt) {
-  var millis = dt * 1
-  var seconds = Math.floor(millis / 1000)
-  var nanos = (millis - (seconds * 1000)) * 1E6
-
-  if (nanos || seconds > 0xFFFFFFFF) {
-    // Timestamp64
-    const encoded = Buffer.allocUnsafe(9)
-    encoded[0] = -1
-
-    var upperNanos = ((nanos * 4))
-    var upperSeconds = seconds / Math.pow(2, 32)
-    var upper = (upperNanos + upperSeconds) & 0xFFFFFFFF
-    var lower = seconds & 0xFFFFFFFF
-
-    encoded.writeInt32BE(upper, 1)
-    encoded.writeInt32BE(lower, 5)
-    return encoded
-  } else {
-    // Timestamp32
-    const encoded = Buffer.allocUnsafe(5)
-    encoded[0] = -1
-    encoded.writeUInt32BE(Math.floor(millis / 1000), 1)
-    return encoded
-  }
 }
 
 function encodeNumber (obj, options) {
